added print messages to help me follow flow of execution

+++ Address Book/MasterViewController.swift
@@ -12,6 +12,7 @@ import CoreData
 class MasterViewController: UITableViewController,NSFetchedResultsControllerDelegate,  AddEditTableViewControllerDelegate, DetailViewControllerDelegate
 {
 
+let showMe = false
   //NSFetchedResultsController informs MasterViewcontroller if the underlying data has changed i.e. a Contact has been changed
 
   var detailViewController: DetailViewController? = nil
@@ -21,7 +22,7 @@ class MasterViewController: UITableViewController,NSFetchedResultsControllerDele
   override func awakeFromNib() 
   {
      super.awakeFromNib()
-print("Master.awakeFromNib")
+if (showMe) {print("M.awakeFromNib")}
     if UIDevice.current.userInterfaceIdiom == .pad
     {
        self.clearsSelectionOnViewWillAppear = false
@@ -35,7 +36,7 @@ print("Master.awakeFromNib")
   {
     super.viewWillAppear(animated)
     //self.clearsSelectionOnViewWillAppear = self.splitViewController!.isCollapsed
-print("Master.viewWillAppear")
+if (showMe) {print("Master.viewWillAppear")}
     displayFirstContactOrInstructions()
   }
 
@@ -43,16 +44,16 @@ print("Master.viewWillAppear")
   // select first contact or display InstructionsViewController
   func displayFirstContactOrInstructions()
   {
-print("Master.displayFirstContactOrInstructions")
+if (showMe) {print("M.displayFirstContactOrInstructions")}
     if let splitViewController = self.splitViewController
     {
-print("Master.splitViewController")
+if (showMe) {print("M.splitViewController")}
       if !splitViewController.isCollapsed
       { // select and display first contact if there is one
-print("Master.splitViewController.isCollapsed")
+if (showMe) {print("M.splitVC.isCollapsed")}
         if self.tableView.numberOfRows(inSection: 0) > 0 
         {
-print("Master.tableView.numberOfRows")
+if (showMe) {print("M.tableView.numberOfRows")}
           let indexPath = NSIndexPath(row: 0, section: 0)
           self.tableView.selectRow(at: indexPath as IndexPath, 
                animated: false, 
@@ -61,7 +62,7 @@ print("Master.tableView.numberOfRows")
         }
         else
         { // display InstructionsViewController
-print("Master.performSegue.showInstructions")
+if (showMe) {print("M.performSegue.showInstructions")}
           self.performSegue(withIdentifier: "showInstructions", sender: self)
         }
       }
@@ -79,10 +80,10 @@ print("Master.performSegue.showInstructions")
      let addButton = UIBarButtonItem(barButtonSystemItem: .add, target: self, action: #selector(insertNewObject(_:)))
      self.navigationItem.rightBarButtonItem = addButton
      */
-print("Master.viewDidLoad")
+if (showMe) {print("M.viewDidLoad")}
     if let split = self.splitViewController
     {
-print("Master.splitViewController2")
+if (showMe) {print("M.splitViewController2")}
       let controllers = split.viewControllers
       self.detailViewController = 
             (controllers[controllers.count-1] as! UINavigationController).topViewController as? DetailViewController
@@ -93,13 +94,13 @@ print("Master.splitViewController2")
 
    override func prepare(for segue: UIStoryboardSegue, sender: Any?)
    {
-print("Master.prepareForSegue")
+if (showMe) {print("M.prepareForSegue")}
       if segue.identifier == "showContactDetail"
       {
-print("Master.prepareForSegue.showContactDetail")
+if (showMe) {print("M.prepareForSegue.showContactDetail")}
          if let indexPath = self.tableView.indexPathForSelectedRow
          {
-print("Master.indexPath")
+if (showMe) {print("M.indexPath")}
             // get Contact for selected row
             let selectedContact = self.fetchedResultsController.object(at: indexPath) as Contact
 
@@ -113,7 +114,7 @@ print("Master.indexPath")
       }
       else if segue.identifier == "showAddContact"
       {
-print("Master.prepareForSegue.showAddContact")
+if (showMe) {print("M.prepareForSegue.showAddContact")}
          // create a contact object that is not yet managed
          let entity = self.fetchedResultsController.fetchRequest.entity!
          let newContact = Contact(entity: entity, 
@@ -135,7 +136,7 @@ print("Master.prepareForSegue.showAddContact")
    // called by AddEditViewController after a contact is added
    func didSaveContact(controller: AddEditTableViewController)
    {
-print("Master.didSaveContact")
+if (showMe) {print("M.didSaveContact")}
     // get NSManagedObjectContext and insert new contact into it
     let context = self.fetchedResultsController.managedObjectContext
     context.insert(controller.contact!)
@@ -149,18 +150,18 @@ if (nserror != nil)
       displayError(error: nserror, title: "Error Saving Data in Master at 1\n",
                    message: "Unable to save contact")
     } else { // if no error, display new contact details
-print("Master.newContactSaved")
+if (showMe) {print("M.newContactSaved")}
       let sectionInfo =
       self.fetchedResultsController.sections![0] as NSFetchedResultsSectionInfo
 //  if let row = FIND(sectionInfo.objects as [NSManagedObject], controller.contact!) {
 let row = sectionInfo.numberOfObjects-1 // minus 1 ????
 if !(row < 0)
 {
-print("Master.row = 0")
+if (showMe) {print("M.row = 0")}
         let path = NSIndexPath(row: row, section: 0)
         tableView.selectRow(at: path as IndexPath, animated: true, scrollPosition: .middle)
 }
-print("Master.performSegue.showContactDetail")
+if (showMe) {print("M.performSegue.showContactDetail")}
         performSegue(withIdentifier: "showContactDetail",
                                    sender: nil)
       }
@@ -169,7 +170,7 @@ print("Master.performSegue.showContactDetail")
    // called by DetailViewController after a contact is edited
    func didEditContact(controller: DetailViewController)
    {
-print("Master.didEditContact")
+if (showMe) {print("M.didEditContact")}
       let context = self.fetchedResultsController.managedObjectContext
 
 //      var error: NSError? = nil
@@ -177,12 +178,12 @@ print("Master.didEditContact")
 //  {
       do
       {
-print("Master.do try context.save")
+if (showMe) {print("M.do try context.save")}
          try context.save()
       }
       catch
       {
-print("Master.do catch")
+if (showMe) {print("M.do catch")}
         // Replace this implementation with code to handle the error appropriately.
         // fatalError() causes the application to generate a crash log and terminate. You should not use this function in a shipping application, although it may be useful during development.
         let nserror = error as NSError
@@ -197,7 +198,7 @@ print("Master.do catch")
    // indicate that an error occurred hen saving database changes
    func displayError(error: NSError?, title: String, message: String)
    {
-print("Master.displayError")
+if (showMe) {print("M.displayError")}
        let alertController = UIAlertController(title: title,
                  message: String(format:  "%@\nError:\(error)\n", message),
                  preferredStyle: UIAlertControllerStyle.alert)
@@ -213,14 +214,14 @@ print("Master.displayError")
 
   override func numberOfSections(in tableView: UITableView) -> Int
   {
-print("Master.numberOfSections")
+if (showMe) {print("M.numberOfSections")}
     return self.fetchedResultsController.sections?.count ?? 0
   }
 
   // callback that returns  number of rows in the UITbleView
   override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int
   {
-print("Master.tableView.numberOfRowsInSection")
+if (showMe) {print("M.tableView.numberOfRowsInSection")}
      let sectionInfo = self.fetchedResultsController.sections![section] as NSFetchedResultsSectionInfo
      return sectionInfo.numberOfObjects
   }
@@ -228,7 +229,7 @@ print("Master.tableView.numberOfRowsInSection")
   // callback that returns a configured cell for the given NSIndexPath
   override func tableView(_ tableView: UITableView, cellForRowAt indexPath:  IndexPath) -> UITableViewCell
   {
-print("Master.tableView.cellForRowAt")
+if (showMe) {print("M.tableView.cellForRowAt")}
     let cell = tableView.dequeueReusableCell(withIdentifier: "Cell", for: indexPath)
     let event = self.fetchedResultsController.object(at: indexPath)
     self.configureCell(cell, withEvent: event)
@@ -238,7 +239,7 @@ print("Master.tableView.cellForRowAt")
   // callback that returns whether a cell is editable
   override func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool 
   {
-print("Master.tableView.canEditRowAt")
+if (showMe) {print("M.tableView.canEditRowAt")}
     // Return false if you do not want the specified item to be editable.
     return true
   }
@@ -246,21 +247,21 @@ print("Master.tableView.canEditRowAt")
   // callback that deletes a row from the UITableView
   override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCellEditingStyle, forRowAt indexPath: IndexPath) 
   {
-print("Master.tableView.commitEditingStyle")
+if (showMe) {print("M.tableView.commitEditingStyle")}
     if editingStyle == .delete 
     {
-print("Master.editingStyle")
+if (showMe) {print("M.editingStyle")}
       let context = self.fetchedResultsController.managedObjectContext
       context.delete(self.fetchedResultsController.object(at: indexPath))
             
       do
       {
-print("Master.do2.context.save")
+if (showMe) {print("M.do2.context.save")}
         try context.save()
       }
       catch
       {
-print("Master.do2 catch")
+if (showMe) {print("M.do2 catch")}
          // Replace this implementation with code to handle the error appropriately.
          // fatalError() causes the application to generate a crash log and terminate. You should not use this function in a shipping application, although it may be useful during development.
          let nserror = error as NSError
@@ -273,16 +274,16 @@ print("Master.do2 catch")
 
   func configureCell(_ cell: UITableViewCell, withEvent event: Contact)
   {
-print("Master.configureCell.withEvent")
+if (showMe) {print("M.configureCell.withEvent")}
     if (managedObjectContext != nil) // ??????
     {
-print("Master.configureCell.gotData")
+if (showMe) {print("M.configureCell.gotData")}
        cell.textLabel?.text = event.firstname?.description
        cell.detailTextLabel?.text = event.lastname?.description
     }
     else
     {
-print("Master.configureCell.gotNoData")
+if (showMe) {print("M.configureCell.gotNoData")}
        cell.textLabel?.text = "Last Name"
 //    cell.textLabel!.text = event.timestamp!.description
 //    cell.textLabel!.text = event.lastname
@@ -296,10 +297,10 @@ print("Master.configureCell.gotNoData")
 
   var fetchedResultsController: NSFetchedResultsController<Contact>
   {
-print("Master.fetchedResultsController")
+if (showMe) {print("M.fetchedResultsController")}
     if _fetchedResultsController != nil
     {
-print("Master.fetchedResultscontroller.nil")
+if (showMe) {print("M.fetchedResultscontroller.nil")}
       return _fetchedResultsController!
     }
       
@@ -326,12 +327,12 @@ print("Master.fetchedResultscontroller.nil")
       
     do 
     {
-print("Master.do3 try.fetchedResultsController")
+if (showMe) {print("M.do3 try.fetchedResultsController")}
       try _fetchedResultsController!.performFetch()
     } 
     catch 
     {
-print("Master.do3 catch.fetchedResultsController")
+if (showMe) {print("M.do3 catch.fetchedResultsController")}
       // Replace this implementation with code to handle the error appropriately.
       // fatalError() causes the application to generate a crash log and terminate. You should not use this function in a shipping application, although it may be useful during development.
       let nserror = error as NSError
@@ -346,14 +347,13 @@ print("Master.do3 catch.fetchedResultsController")
 
   func controllerWillChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) 
   {
-print("Master.controllerWillChangeContect")
+if (showMe) {print("M.controllerWillChangeContect")}
     self.tableView.beginUpdates()
   }
 
-
   func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>, didChange sectionInfo: NSFetchedResultsSectionInfo, atSectionIndex sectionIndex: Int, for type: NSFetchedResultsChangeType) 
   {
-print("Master.controller.didChange sectionInfo")
+if (showMe) {print("M.controller.didChange sectionInfo")}
       switch type 
       {
         case .insert:
@@ -367,7 +367,7 @@ print("Master.controller.didChange sectionInfo")
 
   func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>, didChange anObject: Any, at indexPath: IndexPath?, for type: NSFetchedResultsChangeType, newIndexPath: IndexPath?) 
   {
-print("Master.controller.didChange anObject")
+if (showMe) {print("M.controller.didChange anObject")}
       switch type 
       {
         case .insert:
@@ -383,7 +383,7 @@ print("Master.controller.didChange anObject")
 
   func controllerDidChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) 
   {
-print("Master.controllerDidChangeContent.endUpdates")
+if (showMe) {print("M.controllerDidChangeContent.endUpdates")}
       self.tableView.endUpdates()
   }
